Backend-FastAPI/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
google-cloud >= 0.34.0
google-cloud-texttospeech >= 2.14.1
google-cloud-speech >= 2.21.0
google-cloud-translate >= 3.12.0
pyaudio >= 0.2.13
websockets >= 11.0.3
six >= 1.16.0